v.1.5.6
this is the Version 1.5.6. from [https://gateway.threema.ch/en/developer/sdk-php Threema GmbH] as they do not maintain their repo anymore (but refer to it in terms of documentation) So unfortunately there is only very little documentation available and we all have to dig through their php code. Luckily it is quite well written and readable